Listar y Editar Cats

+++ repo_tpfinal/Negocio/NegocioCategoria.cs
@@ -12,20 +12,20 @@ namespace Negocio
     public class NegocioCategoria
     {
 
-        public bool eliminarCategoria(int id)
-        {
+        //public bool eliminarCategoria(int id)
+        //{
             //Validar id existente 
-            DaoCategoria dao = new DaoCategoria();
-            Categorias cat = new Categorias();
+            //DaoCategoria dao = new DaoCategoria();
+            //Categorias cat = new Categorias();
             //cat.setIdCategoria(id);
-            cat.Id_categoria = id;
+            //cat.Id_categoria = id;
 
-            int op = dao.eliminarCategoria(cat);
-            if (op == 1)
-                return true;
-            else
-                return false;
-        }
+            //int op = dao.eliminarCategoria(cat);
+            //if (op == 1)
+              //  return true;
+            //else
+              //  return false;
+        //}
 
         public bool agregarCategoria(String nombre)
         {
@@ -46,18 +46,49 @@ namespace Negocio
             else
                 return false;
         }
-        public bool eliminarMarcas_neg(String nombre)
+
+        public bool actualizarCategoria(Categorias cat)
         {
+            int cantFilas = 0;
+            //Categorias cat = new Categorias;
+            DaoCategoria dao = new DaoCategoria();
+
+            cantFilas = dao.actualizarCategoria(cat);
+
+            if (cantFilas == 1)
+            {
+                return true;
+            }
+            else
+            {
+                return false;
+            }
+                
+
+        }
+
+        
+
+        public bool eliminarCategoria(String nombre)
+        {
+            int op = 0;
             //Validar id existente 
             DaoCategoria dao = new DaoCategoria();
             Categorias cat = new Categorias();
             cat.Nombre1 = nombre;
-            int op = dao.eliminarCategoria(cat);
+            if (dao.existeCategoría(cat) == true)
+            {
+                op = dao.eliminarCategoria(cat);
+            }
+            
             if (op == 1)
                 return true;
             else
                 return false;
         }
+
+
+
         public DataTable TodasLasCategorias()
         {
             DaoCategoria dp = new DaoCategoria();
@@ -68,6 +99,9 @@ namespace Negocio
             DaoCategoria dm = new DaoCategoria();
             return dm.ObtenerCategorias();
         }
-        
+
+        //////////////////////
+
+
     }
 }
